+++ devkit/loadtesting-k6/src/rpc-server/index.ts
@@ -1,6 +1,8 @@
-import { getOption } from "./optionBuilder.js";
-import { JsonRpcMethodCounters, buildWallet, instantiateCounters } from "./requestBuilder.js";
+import { getOption } from "./utils/test-option.js";
+import { instantiatePayloadCounters } from "./utils/rpc-payload.js";
 import { OptionTypes, WalletTypes } from "./types.js";
+import { JsonRpcMethodCounters } from "./wallet/mock-wallet.js";
+import { buildWallet } from "./wallet/index.js";
 
 // get values from env
 const url = __ENV.RPC_SERVER_URL;
@@ -11,21 +13,10 @@ const walletType = __ENV.WALLET_TYPE as WalletTypes;
 export const options = getOption(testType);
 
 // create metrics counter
-const counters: JsonRpcMethodCounters = instantiateCounters();
+const counters: JsonRpcMethodCounters = instantiatePayloadCounters();
 
 // Simulated user behavior
 export default function(): void {
   const wallet = buildWallet(walletType, url);
   wallet.sendRpcCalls(counters);
-  // const [url, body, params] = getRequest(requestType, count);
-  // const res = http.post(url, body, params);
-
-  // // Validate response status
-  // check(res, {
-  //   "status was 200": (r) => r.status == 200,
-  //   "verify resp": (r) =>
-  //     typeof r.body == "string" &&
-  //     r.body.includes("jsonrpc") &&
-  //     !r.body.includes("error"),
-  // });
 }
